+++ cli/modules/source/index.js
@@ -6,14 +6,13 @@ const assert = require('assert').strict
 const inquirer = require('inquirer')
 const runner = require('jscodeshift/src/Runner')
 const hq = require('../../src')
-const { makeObjectBullets } = require('../common')
+const { showConfig, checkPaths, makeItemsBullets, makeObjectBullets, makePathsBullets } = require('../services/common')
+const { getAliases, numAliases, saveSettings } = require('../services/config')
 const { inspect } = require('../utils')
 const { getLongestStringLength, makeHeader } = require('../utils/text')
-const { getAliases, numAliases, saveSettings } = require('../utils/config')
 const { getPathInfo, cleanPathsInfo } = require('../utils/paths')
 const { makeChoices } = require('../utils/inquirer')
 const { para } = require('../utils/text')
-const { showConfig, checkPaths, makeItemsBullets, makePathsBullets } = require('../common')
 const { TransformMode } = require('./paths')
 const stats = require('./stats')
