+++ cli/modules/source/paths.js
