+++ cli/modules/source/stats.js
